backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["artem <artem2004920@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.44"
fastapi = "^0.121.2"
pydantic-settings = "^2.12.0"
structlog = "^25.5.0"
asyncpg = "^0.30.0"
uvicorn = "^0.38.0"
passlib = "^1.7.4"
python-jose = "^3.5.0"
sqladmin = "^0.21.0"
itsdangerous = "^2.2.0"
langchain-text-splitters = "*"
pdfminer-six = "20231228"
httpx = "^0.27.2"
qdrant-client = "^1.9.1"
minio = "^7.2.7"
python-docx = "^1.1.2"
python-pptx = "^0.6.23"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
